[MOVE/#14] first->select 이름 변경
