--- feature/evaluate/src/main/java/com/my/version/feature/evaluate/first/EvaluationFirstViewModel.kt
@@ -1,10 +1,0 @@
-package com.my.version.feature.evaluate.first
-
-import androidx.lifecycle.ViewModel
-import javax.inject.Inject
-
-class EvaluationFirstViewModel @Inject constructor(
-
-): ViewModel() {
-
-}
