+++ feature/evaluate/src/main/java/com/my/version/feature/evaluate/select/EvaluationSelectViewModel.kt
@@ -1,0 +1,10 @@
+package com.my.version.feature.evaluate.select
+
+import androidx.lifecycle.ViewModel
+import javax.inject.Inject
+
+class EvaluationSelectViewModel @Inject constructor(
+
+): ViewModel() {
+
+}
