+++ feature/evaluate/src/main/java/com/my/version/feature/evaluate/select/navigation/EvaluationSelectNavigation.kt
@@ -1,4 +1,4 @@
-package com.my.version.feature.evaluate.first.navigation
+package com.my.version.feature.evaluate.select.navigation
 
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavController
@@ -6,21 +6,21 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.my.version.core.common.navigation.Route
-import com.my.version.feature.evaluate.first.EvaluationFirstRoute
+import com.my.version.feature.evaluate.select.EvaluationSelectRoute
 import com.my.version.feature.evaluate.main.navigation.Evaluation
 import kotlinx.serialization.Serializable
 
-fun NavController.navigateToEvaluationFirst(navOptions: NavOptions) = navigate(Evaluation, navOptions)
+fun NavController.navigateToEvaluationSelect(navOptions: NavOptions) = navigate(Evaluation, navOptions)
 
-fun NavGraphBuilder.evaluationFirstScreen(
+fun NavGraphBuilder.evaluationSelectScreen(
     modifier: Modifier
 ) {
-    composable<EvaluationFirst>{
-        EvaluationFirstRoute(
+    composable<EvaluationSelect>{
+        EvaluationSelectRoute(
             modifier = modifier
         )
     }
 }
 
 @Serializable
-data object EvaluationFirst: Route
+data object EvaluationSelect: Route
