+++ feature/main/src/main/java/com/my/version/feature/main/MainScreen.kt
@@ -28,7 +28,7 @@ import com.my.version.feature.auth.signup.navigation.signUpScreen
 import com.my.version.feature.cover.navigation.coverFirstScreen
 import com.my.version.feature.cover.navigation.coverScreen
 import com.my.version.feature.cover.navigation.coverSecondScreen
-import com.my.version.feature.evaluate.first.navigation.evaluationFirstScreen
+import com.my.version.feature.evaluate.select.navigation.evaluationSelectScreen
 import com.my.version.feature.evaluate.main.navigation.evaluationScreen
 import com.my.version.feature.home.navigation.homeScreen
 import com.my.version.feature.home.navigation.navigateToHome
@@ -89,7 +89,7 @@ fun MainNavHost(
         evaluationScreen(
             modifier = bottomBarModifier
         )
-        evaluationFirstScreen(
+        evaluationSelectScreen(
             modifier = noBottomBarModifier
         )
         signInScreen(
